Add files via upload
This file contains Linear Regression on a Generated Dataset. In this, we learn about primary libraries such as numpy and matplotlib in Python. We learn about how we can generate a dataset using random values. We also learn how do we find the best solution for bias and intercept term with the help of normal equation.
Also, we explore Linear Regression implementation using SKLearn Library.
Thank you and keep learning!

Learnt from and Credits to the Book: Hands-On Machine Learning with Scikit-Learn,